Added multiple new functions for WebScraper.py. Updated the main file and README.MD.
Added ability to detect pop-ups with text inputs, Request URLs, Anchor Links, Age of Domains, if a site has a valid SSL certificate, and the length of a domain.

Updated the main file to format data in the form of a dictionary to be parsed into the model.

Updated the README.md to specify the operating system and software required.